Detect disconnected devices better
for disconnected but present (via dongle?) devices
OpenRazer seems to return battery_level 0 which is not really
helpful.
One device attrbiute I've found that is distinct when
device is connected or not is 'get_idle_time'
(at least on that single mouse I am testing this all on :-) )
so use that to return only connected devices.